// Web/GuessWord.Frontend/src/app/clients/api.client.ts
export class WordDto implements IWordDto {
    id?: number;
    value?: string | undefined;
    language?: number;
    translations?: string[] | undefined;

    constructor(data?: IWordDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.value = _data["value"];
            this.language = _data["language"];
            if (Array.isArray(_data["translations"])) {
                this.translations = [] as any;
                for (let item of _data["translations"])
                    this.translations!.push(item);
            }
        }
    }

    static fromJS(data: any): WordDto {
        data = typeof data === 'object' ? data : {};
        let result = new WordDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["value"] = this.value;
        data["language"] = this.language;
        if (Array.isArray(this.translations)) {
            data["translations"] = [];
            for (let item of this.translations)
                data["translations"].push(item);
        }
        return data;
    }
}

export interface IWordDto {
    id?: number;
    value?: string | undefined;
    language?: number;
    translations?: string[] | undefined;
}